refactor: Separate out functions from generateHtml
- getDecorations
- renderClipboard
- renderClipboardAndError
- renderError
