+++ packages/test-case-component/src/helpers/decorations/index.ts
@@ -1,0 +1,1 @@
+export * from "./getDecorations";
