+++ packages/test-case-component/src/renderHtml/index.ts
@@ -1,0 +1,2 @@
+export * from "./renderClipboard";
+export * from "./renderError";
