+++ packages/test-case-component/src/renderHtml/renderClipboardAndError.ts
@@ -1,0 +1,1 @@
+// Remove the old renderClipboardAndError.ts file since the functions are now in their own files
